refactor: move Affix.tsx and Icon.tsx into the Adornment dir
