--- src/components/TextInput/Affix.tsx
@@ -1,99 +1,0 @@
-import React from 'react';
-import color from 'color';
-import {
-  Text,
-  StyleSheet,
-  StyleProp,
-  ViewStyle,
-  TextStyle,
-  LayoutChangeEvent,
-  Animated,
-} from 'react-native';
-
-import { withTheme } from '../../core/theming';
-import { Theme } from '../../types';
-
-const AFFIX_OFFSET = 12;
-
-type Props = {
-  text: string;
-  onLayout?: (event: LayoutChangeEvent) => void;
-  style?: StyleProp<ViewStyle>;
-  textStyle?: StyleProp<TextStyle>;
-  visible?: Animated.Value;
-  /**
-   * @optional
-   */
-  theme: Theme;
-};
-
-export function renderAffix({
-  affix,
-  side,
-  textStyle,
-  affixTopPosition,
-  onLayout,
-  visible,
-}: {
-  affix: React.ReactNode;
-  side: 'left' | 'right';
-  textStyle: StyleProp<TextStyle>;
-  affixTopPosition: number | null;
-  onLayout?: (event: LayoutChangeEvent) => void;
-  visible?: Animated.Value;
-}): React.ReactNode {
-  //@ts-ignore
-  return React.cloneElement(affix, {
-    style: {
-      top: affixTopPosition,
-      [side]: AFFIX_OFFSET,
-    },
-    textStyle,
-    onLayout,
-    visible,
-  });
-}
-
-class TextInputAffix extends React.Component<Props> {
-  static displayName = 'TextInput.Affix';
-
-  render() {
-    const { text, style, textStyle, onLayout, theme, visible } = this.props;
-
-    const textColor = color(theme.colors.text)
-      .alpha(theme.dark ? 0.7 : 0.54)
-      .rgb()
-      .string();
-
-    return (
-      <Animated.View
-        style={[
-          styles.container,
-          style,
-          {
-            opacity:
-              visible?.interpolate({
-                inputRange: [0, 1],
-                outputRange: [1, 0],
-              }) || 1,
-          },
-        ]}
-        onLayout={onLayout}
-      >
-        <Text style={[{ color: textColor }, textStyle]}>{text}</Text>
-      </Animated.View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    position: 'absolute',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
-
-export default withTheme(TextInputAffix);
-
-export { TextInputAffix };
